refactor in raffle package, sample ErrorConsumer
