+++ docker-compose.yml
@@ -26,7 +26,6 @@ services:
       - mysql
       - rabbitmq
     restart: on-failure:10
-
   mysql:
     image: mysql
     container_name: mysql
@@ -38,7 +37,6 @@ services:
     ports:
       - 3306:3306
     restart: on-failure:10
-
   rabbitmq:
     image: rabbitmq:3.8-management
     container_name: rabbitmq
