+++ src/main/java/com/traanite/plumfish/raffle/infrastructure/PseudoRandomNumberGenerator.java
@@ -1,18 +1,17 @@
 package com.traanite.plumfish.raffle.infrastructure;
 
-import com.traanite.plumfish.commons.events.exception.NullResponseException;
 import com.traanite.plumfish.raffle.model.RandomNumberGenerator;
-import com.traanite.plumfish.raffle.model.RandomNumberGeneratorError;
 import com.traanite.plumfish.raffle.model.Range;
 import lombok.RequiredArgsConstructor;
 
+import java.util.Optional;
 import java.util.concurrent.ThreadLocalRandom;
 
 @RequiredArgsConstructor
 public class PseudoRandomNumberGenerator implements RandomNumberGenerator {
 
     @Override
-    public int randomInt(Range range) throws NullResponseException, RandomNumberGeneratorError {
-        return ThreadLocalRandom.current().nextInt(range.getMin(), range.getMax() + 1);
+    public Optional<Integer> randomInt(Range range) {
+        return Optional.of(ThreadLocalRandom.current().nextInt(range.getMin(), range.getMax()));
     }
 }
