+++ src/main/java/com/traanite/plumfish/raffle/model/PackageDrawer.java
@@ -1,6 +1,5 @@
 package com.traanite.plumfish.raffle.model;
 
-import com.traanite.plumfish.commons.events.exception.NullResponseException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -15,36 +14,22 @@ public class PackageDrawer {
     private final RandomNumberGenerator randomNumberGenerator;
 
     public Optional<RafflePackage> drawPackage() {
-        final Integer drawnNumber = randomIntOrLogError(new Range(1, 10000));
-        final Integer drawnStarNumber = randomIntOrLogError(new Range(1, stars.getData().size()));
-        final Integer drawnThingNumber = randomIntOrLogError(new Range(1, things.getNames().size()));
+        final Optional<Integer> drawnNumber = randomNumberGenerator.randomInt(new Range(1, 10000));
+        final Optional<Integer> drawnStarNumber = randomNumberGenerator.randomInt(new Range(1, stars.getData().size()));
+        final Optional<Integer> drawnThingNumber = randomNumberGenerator.randomInt(new Range(1, things.getNames().size()));
 
-        if (!allNumbersValid(drawnNumber, drawnStarNumber, drawnThingNumber)) {
+        if (!(drawnNumber.isPresent() && drawnStarNumber.isPresent() && drawnThingNumber.isPresent())) {
             return Optional.empty();
         }
 
-        Star drawnStar = starWithHrEqualNumberOrAny(drawnStarNumber);
-        Thing thing = new Thing(things.getNames().get(drawnThingNumber));
+        Star drawnStar = starWithHrEqualNumberOrAny(drawnStarNumber.get());
+        Thing thing = new Thing(things.getNames().get(drawnThingNumber.get()));
 
-        return Optional.of(new RafflePackage(drawnNumber, drawnStar, thing));
+        return Optional.of(new RafflePackage(drawnNumber.get(), drawnStar, thing));
     }
 
     private Star starWithHrEqualNumberOrAny(int drawnStarNumber) {
         return stars.getData().stream().filter(star -> star.getHr() == drawnStarNumber).findFirst()
                 .orElse(stars.getData().stream().findAny().orElseThrow());
     }
-
-    private boolean allNumbersValid(Integer drawnNumber, Integer drawnStarNumber, Integer drawnThingNumber) {
-        return drawnNumber != null && drawnStarNumber != null && drawnThingNumber != null;
-    }
-
-    private Integer randomIntOrLogError(Range range) {
-        try {
-            return randomNumberGenerator.randomInt(range);
-        } catch (RandomNumberGeneratorError | NullResponseException e) {
-            log.error("Couldn't get random number", e);
-            return null;
-        }
-    }
-
 }
